Decorator to middleware
